fix: BigDecimal -> Float (because of using all of 2^(-n))

+++ src/main/java/com/dyinfotech/annualleavebackend/domain/LeaveRequest.java
@@ -1,14 +1,25 @@
 package com.dyinfotech.annualleavebackend.domain;
 
-import jakarta.persistence.*;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.PrePersist;
+import jakarta.persistence.PreUpdate;
+import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "leave_request")
@@ -31,8 +42,9 @@ public class LeaveRequest {
     @Column(name = "end_date", nullable = false)
     private LocalDate endDate;
 
+    // TODO: 반반차 도입시 Decimal(5, 2). 1시간 연차 도입시 Decimal(6, 3)
     @Column(name = "use_days", nullable = false, precision = 4, scale = 1)
-    private BigDecimal useDays;
+    private Float useDays;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "status", nullable = false, length = 10)
@@ -55,7 +67,7 @@ public class LeaveRequest {
     private LocalDateTime updatedAt;
 
     @Builder
-    public LeaveRequest(Employee employee, LocalDate startDate, LocalDate endDate, BigDecimal useDays) {
+    public LeaveRequest(Employee employee, LocalDate startDate, LocalDate endDate, Float useDays) {
         this.employee = employee;
         this.startDate = startDate;
         this.endDate = endDate;
